# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * BidRush Auction Platform API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
  message?: string;
}

export interface RegisterRequest {
  name: string;
  email: string;
  /** @minLength 6 */
  password: string;
  avatar?: string | null;
}

export interface LoginRequest {
  email: string;
  password: string;
}

export interface User {
  id: string;
  name: string;
  email: string;
  avatar?: string | null;
  rating: number;
  totalSales: number;
  totalPurchases: number;
  isAdmin: boolean;
  createdAt: string;
}

export interface AuthResponse {
  token: string;
  user: User;
}

export interface UserProfile {
  id: string;
  name: string;
  avatar?: string | null;
  rating: number;
  totalSales: number;
  totalPurchases: number;
  createdAt: string;
}

export type AuctionStatus = (typeof AuctionStatus)[keyof typeof AuctionStatus];

export const AuctionStatus = {
  live: "live",
  scheduled: "scheduled",
  ended: "ended",
} as const;

export interface Auction {
  id: string;
  title: string;
  description: string;
  images: string[];
  category: string;
  startingPrice: number;
  currentPrice: number;
  minimumIncrement: number;
  status: AuctionStatus;
  startTime: string;
  endTime: string;
  sellerId: string;
  sellerName: string;
  sellerAvatar?: string | null;
  bidCount: number;
  winnerId?: string | null;
  winnerName?: string | null;
  createdAt: string;
}

export interface Bid {
  id: string;
  auctionId: string;
  bidderId: string;
  bidderName: string;
  bidderAvatar?: string | null;
  amount: number;
  createdAt: string;
}

export type AuctionDetail = Auction & {
  recentBids: Bid[];
};

export interface AuctionListResponse {
  auctions: Auction[];
  total: number;
  page: number;
  totalPages: number;
}

export interface CreateAuctionRequest {
  title: string;
  description: string;
  images: string[];
  category: string;
  startingPrice: number;
  minimumIncrement: number;
  startTime: string;
  endTime: string;
}

export type UpdateAuctionRequestStatus =
  (typeof UpdateAuctionRequestStatus)[keyof typeof UpdateAuctionRequestStatus];

export const UpdateAuctionRequestStatus = {
  live: "live",
  scheduled: "scheduled",
  ended: "ended",
} as const;

export interface UpdateAuctionRequest {
  title?: string;
  description?: string;
  images?: string[];
  status?: UpdateAuctionRequestStatus;
}

export type BidWithAuction = Bid & {
  auction: Auction;
};

export interface PlaceBidRequest {
  amount: number;
}

export type NotificationType =
  (typeof NotificationType)[keyof typeof NotificationType];

export const NotificationType = {
  outbid: "outbid",
  auction_ending: "auction_ending",
  auction_won: "auction_won",
  auction_lost: "auction_lost",
  new_bid: "new_bid",
} as const;

export interface Notification {
  id: string;
  userId: string;
  type: NotificationType;
  title: string;
  message: string;
  auctionId?: string | null;
  read: boolean;
  createdAt: string;
}

export type ListAuctionsParams = {
  status?: ListAuctionsStatus;
  category?: string;
  search?: string;
  minPrice?: number;
  maxPrice?: number;
  page?: number;
  limit?: number;
};

export type ListAuctionsStatus =
  (typeof ListAuctionsStatus)[keyof typeof ListAuctionsStatus];

export const ListAuctionsStatus = {
  live: "live",
  scheduled: "scheduled",
  ended: "ended",
} as const;

export type GetAuctionBids200 = {
  bids: Bid[];
};

export type GetMyAuctions200 = {
  auctions: Auction[];
};

export type GetMyBids200 = {
  bids: BidWithAuction[];
};

export type GetMyWonAuctions200 = {
  auctions: Auction[];
};

export type GetNotifications200 = {
  notifications: Notification[];
};

export type AdminListUsers200 = {
  users: User[];
};