# lib/api-zod/src/generated/types/auction.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * BidRush Auction Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { AuctionStatus } from "./auctionStatus";

export interface Auction {
  id: string;
  title: string;
  description: string;
  images: string[];
  category: string;
  startingPrice: number;
  currentPrice: number;
  minimumIncrement: number;
  status: AuctionStatus;
  startTime: Date;
  endTime: Date;
  sellerId: string;
  sellerName: string;
  sellerAvatar?: string | null;
  bidCount: number;
  winnerId?: string | null;
  winnerName?: string | null;
  createdAt: Date;
}