# lib/api-zod/src/generated/types/createAuctionRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * BidRush Auction Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateAuctionRequest {
  title: string;
  description: string;
  images: string[];
  category: string;
  startingPrice: number;
  minimumIncrement: number;
  startTime: Date;
  endTime: Date;
}