# lib/api-zod/src/generated/types/updateAuctionRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * BidRush Auction Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { UpdateAuctionRequestStatus } from "./updateAuctionRequestStatus";

export interface UpdateAuctionRequest {
  title?: string;
  description?: string;
  images?: string[];
  status?: UpdateAuctionRequestStatus;
}