# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * BidRush Auction Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface User {
  id: string;
  name: string;
  email: string;
  avatar?: string | null;
  rating: number;
  totalSales: number;
  totalPurchases: number;
  isAdmin: boolean;
  createdAt: Date;
}